Confirm Delete
In the updated code snippet, the showDeleteStudentConfirmation method is used to display the confirmation message to the user by returning the confirm-delete view template. The id path variable is added to the model to set the th:action attribute of the form in the confirm-delete.html page.

The deleteStudentInformation method is updated to handle the confirmation message by checking the value of the confirm parameter. If the value is yes, the student information is deleted from the database else redirected to list page

+++ src/main/java/com/manage/student/controller/StudentController.java
@@ -1,4 +1,5 @@
 package com.manage.student.controller;
+
 import com.manage.student.model.Student;
 import com.manage.student.repository.StudentRepository;
 import jakarta.validation.Valid;
@@ -36,7 +37,7 @@ public class StudentController {
             bindingResult.addError(new FieldError("student","course","Course name should not be blank."));
         }
         if(student.getAge() < 10){
-            bindingResult.addError(new FieldError("student","age","Enter correct age."));
+            bindingResult.addError(new FieldError("student","age","Enter your correct age."));
         }
         if(bindingResult.hasErrors()){
             return "add-student";
@@ -45,15 +46,33 @@ public class StudentController {
         model.addAttribute("students", studentRepository.findAll());
         return "list-students";
     }
+
     @GetMapping("/delete/{id}")
-    public String deleteStudentInformation(@PathVariable("id")int id, Model model){
+    public String showDeleteStudentConfirmation(@PathVariable("id") int id, Model model) {
         Student student = studentRepository.findById(id)
-                .orElseThrow(()-> new IllegalArgumentException("Invalid Student ID" + id));
-        studentRepository.delete(student);
+                .orElseThrow(() -> new IllegalArgumentException("Invalid Student ID " + id));
+        model.addAttribute("id", id);
+        return "confirm-delete";
+}
+
+    @PostMapping("/delete/{id}")
+    public String deleteStudentInformation(@PathVariable("id") int id, @RequestParam("confirm") String confirm, Model model) {
+        if (confirm.equals("yes")) {
+            Student student = studentRepository.findById(id)
+                    .orElseThrow(() -> new IllegalArgumentException("Invalid Student ID " + id));
+            studentRepository.delete(student);
+        }
         model.addAttribute("students", studentRepository.findAll());
         return "list-students";
     }
 
+    /**
+     * Update function
+     * @GetMapping - GET request at the endpoint "/update/{id}".
+     * @PathVariable - The {id} path variable is extracted from the
+     * endpoint URL and used to identify the student to be updated.
+     * Model object which is used to pass data to the view.
+     */
     @GetMapping("/update/{id}")
     public String moveToUpdateStudentPage(@PathVariable("id")int id, Model model){
         Student student = studentRepository.findById(id)
@@ -62,6 +81,14 @@ public class StudentController {
         return "edit-student";
     }
 
+    /**
+     * Save-Update function
+     * @PostMapping - HTTP POST request at the endpoint "/update/{id}".
+     * @PathVariable - The {id} path variable is extracted from the
+     * endpoint URL and used to identify the student to be updated.
+     * @Valid - the object will be validated based on the validation annotations in the Student class.
+     * BindingResult - used to collect any validation errors that occur during the validation process.
+     */
     @PostMapping("/update/{id}")
     public String updateStudent(@PathVariable("id")int id,@Valid Student student,BindingResult bindingResult,Model model){
         //name validation
@@ -75,9 +102,11 @@ public class StudentController {
         if(student.getCourse().isBlank()){
             bindingResult.addError(new FieldError("student","course","Course name should not be blank."));
         }
+
         if(student.getAge() < 10){
-            bindingResult.addError(new FieldError("student","age","Enter correct age."));
+                bindingResult.addError(new FieldError("student","age","Enter your correct age."));
         }
+
         if(bindingResult.hasErrors()){
             return "edit-student";
         }
@@ -86,7 +115,7 @@ public class StudentController {
         return "list-students";
     }
 
-    // method to handle all Whitelabel errors
+    // method to handle all exceptions and call the error.html
     @ExceptionHandler(Throwable.class)
     public String handleException(Model model, Exception ex) {
         model.addAttribute("errorMessage", ex.getMessage());
